realtime/collectors-primary/requirements.txt
# Collectors-Primary Service
# ──────────────────────────

websockets>=12.0
aiohttp>=3.9
redis[hiredis]>=5.0
python-dotenv>=1.0
